feat!: move kernel MMR position to `u64` (#5956)
Description
---
Switches the handling of kernel MMR positions from `u32` to `u64`.

Closes #5924.

Motivation and Context
---
Internally, MMR positions are handled as either `usize` or `u64`.
However, they are eventually parsed and handled as `u32` for storage.
While it doesn't address the internal `usize` handling, it switches
storage operations to be `u64`.

How Has This Been Tested?
---
Existing tests pass.

What process can a PR reviewer use to test or verify this change?
---
Assert that the logic, especially around LMDB operations, is correct.

BREAKING CHANGE: Modifies the handling of kernel MMR data in LMDB.

+++ base_layer/core/src/chain_storage/async_db.rs
@@ -330,7 +330,7 @@ impl<'a, B: BlockchainBackend + 'static> AsyncDbTransaction<'a, B> {
         &mut self,
         kernel: TransactionKernel,
         header_hash: HashOutput,
-        mmr_position: u32,
+        mmr_position: u64,
     ) -> &mut Self {
         self.transaction.insert_kernel(kernel, header_hash, mmr_position);
         self
